+++ internal/business/notifications/store.go
@@ -15,7 +15,7 @@ func NewStore(db *sql.DB) *NotificationStore {
 	return &NotificationStore{db}
 }
 
-func (h *NotificationStore) Create(email string, timestamp time.Time, weekdays internal.WeekdaysSum) (int64, error) {
+func (h *NotificationStore) Create(email string, timestamp time.Time, weekdays internal.WeekdaysSum) (int, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
 	defer cancel()
 
@@ -29,5 +29,5 @@ func (h *NotificationStore) Create(email string, timestamp time.Time, weekdays i
 	if err != nil {
 		return 0, err
 	}
-	return insertId, nil
+	return int(insertId), nil
 }
